requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.1.78
scikit-learn==1.3.0